pyproject.toml
[project]
name = "mlops-lab1-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.9.0",
    "click>=8.3.0",
    "fastapi>=0.119.1",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "mlflow>=3.7.0",
    "numpy>=2.3.5",
    "onnx>=1.20.0",
    "onnxruntime>=1.23.2",
    "onnxscript>=0.5.6",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pylint>=4.0.2",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "python-multipart>=0.0.20",
    "scipy>=1.16.3",
    "setuptools>=80.9.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "uvicorn>=0.38.0",
]
